requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
cryptography==2.9.2
cycler==0.10.0
defusedxml==0.6.0
Django==3.0.4
django-cors-headers==3.7.0
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-rest-knox==4.1.0
django-webpack-loader==1.1.0
djangorestframework==3.11.0
graphics.py==5.0.1.post1
idna==2.9
oauthlib==3.1.0
Pillow==9.1.0
pycodestyle==2.6.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
stripe==2.27.0
urllib3==1.25.9
